Brain_Tumor_Analysis/backend/code/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
lime==0.2.0.1
matplotlib==3.8.4
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
shap==0.45.0
tensorflow==2.15.0
tensorflow_macos==2.15.0
